+++ platform/darwin/test/MGLPredicateTests.mm
@@ -1,7 +1,7 @@
 #import <XCTest/XCTest.h>
 #import <Mapbox/Mapbox.h>
 
-#import "NSPredicate+MGLAdditions.h"
+#import "NSPredicate+MGLPrivateAdditions.h"
 #import "MGLValueEvaluator.h"
 
 namespace mbgl {
@@ -242,48 +242,48 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"x == YES"];
         NSArray *jsonExpression = @[@"==", @[@"get", @"x"],  @YES];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSNumber') < 5"];
         NSArray *jsonExpression = @[@"<", @[@"to-number", @[@"get", @"x"]], @5];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSNumber') > 5"];
         NSArray *jsonExpression = @[@">", @[@"to-number", @[@"get", @"x"]], @5];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSNumber') <= 5"];
         NSArray *jsonExpression = @[@"<=", @[@"to-number", @[@"get", @"x"]], @5];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSNumber') >= 5"];
         NSArray *jsonExpression = @[@">=", @[@"to-number", @[@"get", @"x"]], @5];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSString') > 'value'"];
         NSArray *jsonExpression = @[@">",  @[@"to-string", @[@"get", @"x"]], @"value"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
@@ -360,48 +360,48 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(a, 'NSString') < 'b'"];
         NSArray *jsonExpression = @[@"<", @[@"to-string", @[@"get", @"a"]], @"b"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(a, 'NSString') <= 'b'"];
         NSArray *jsonExpression = @[@"<=", @[@"to-string", @[@"get", @"a"]], @"b"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(a, 'NSString') > 'b'"];
         NSArray *jsonExpression = @[@">", @[@"to-string", @[@"get", @"a"]], @"b"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(a, 'NSString') >= 'b'"];
         NSArray *jsonExpression = @[@">=", @[@"to-string", @[@"get", @"a"]], @"b"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(a, 'NSString') BETWEEN {'b', 'z'}"];
         NSArray *jsonExpression =@[@"all", @[@"<=", @"b", @[@"to-string", @[@"get", @"a"]]], @[@"<=", @[@"to-string", @[@"get", @"a"]], @"z"]];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSExpression *limits = [NSExpression expressionForAggregate:@[[NSExpression expressionForConstantValue:@10], [NSExpression expressionForConstantValue:@100]]];
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSNumber') BETWEEN %@", limits];
         NSArray *jsonExpression = @[@"all", @[@">=", @[@"to-number", @[@"get", @"x"]], @10], @[@"<=", @[@"to-number", @[@"get", @"x"]], @100]];
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
@@ -409,24 +409,24 @@ namespace mbgl {
         NSExpression *limits = [NSExpression expressionForAggregate:@[[NSExpression expressionForConstantValue:@10], [NSExpression expressionForConstantValue:@100]]];
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSNumber') BETWEEN %@", limits];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, expected);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:expected], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:expected]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:expected], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:expected]
                           mustRoundTrip:NO];
     }
     {
         NSArray *expected = @[@"all", @[@"<=", @10, @[@"to-number", @[@"get", @"x"]]], @[@">=", @100, @[@"to-number", @[@"get", @"x"]]]];
         NSExpression *limits = [NSExpression expressionForAggregate:@[[NSExpression expressionForConstantValue:@10], [NSExpression expressionForConstantValue:@100]]];
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSNumber') BETWEEN %@", limits];
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:expected], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:expected]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:expected], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:expected]
                           mustRoundTrip:NO];
     }
     {
         NSArray *expected = @[@"all", @[@">=", @[@"to-number", @[@"get", @"x"]], @10], @[@">=", @100, @[@"to-number", @[@"get", @"x"]]]];
         NSExpression *limits = [NSExpression expressionForAggregate:@[[NSExpression expressionForConstantValue:@10], [NSExpression expressionForConstantValue:@100]]];
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"CAST(x, 'NSNumber') BETWEEN %@", limits];
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:expected], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:expected]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:expected], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:expected]
                           mustRoundTrip:NO];
     }
     {
@@ -434,7 +434,7 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"$featureIdentifier IN { 6, 5, 4, 3}"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, expected);
         NSPredicate *predicateAfter = [NSPredicate predicateWithFormat:@"MGL_MATCH(CAST($featureIdentifier, 'NSNumber'), 3, YES, 4, YES, 5, YES, 6, YES, NO) == YES"];
-        auto forwardFilter = [NSPredicate mgl_predicateWithJSONObject:expected].mgl_filter;
+        auto forwardFilter = [NSPredicate predicateWithMGLJSONObject:expected].mgl_filter;
         NSPredicate *forwardPredicateAfter = [NSPredicate mgl_predicateWithFilter:forwardFilter];
         XCTAssertEqualObjects(predicateAfter, forwardPredicateAfter);
     }
@@ -443,7 +443,7 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"NOT x IN { 6, 5, 4, 3}"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, expected);
         NSPredicate *predicateAfter = [NSPredicate predicateWithFormat:@"NOT MGL_MATCH(CAST(x, 'NSNumber'), 3, YES, 4, YES, 5, YES, 6, YES, NO) == YES"];
-        auto forwardFilter = [NSPredicate mgl_predicateWithJSONObject:expected].mgl_filter;
+        auto forwardFilter = [NSPredicate predicateWithMGLJSONObject:expected].mgl_filter;
         NSPredicate *forwardPredicateAfter = [NSPredicate mgl_predicateWithFilter:forwardFilter];
         XCTAssertEqualObjects(predicateAfter, forwardPredicateAfter);
     }
@@ -452,7 +452,7 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"a IN { 'b', 'c' }"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, expected);
         NSPredicate *predicateAfter = [NSPredicate predicateWithFormat:@"MGL_MATCH(CAST(a, 'NSString'), 'b', YES, 'c', YES, NO) == YES"];
-        auto forwardFilter = [NSPredicate mgl_predicateWithJSONObject:expected].mgl_filter;
+        auto forwardFilter = [NSPredicate predicateWithMGLJSONObject:expected].mgl_filter;
         NSPredicate *forwardPredicateAfter = [NSPredicate mgl_predicateWithFilter:forwardFilter];
         XCTAssertEqualObjects(predicateAfter, forwardPredicateAfter);
     }
@@ -461,7 +461,7 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"%@ IN %@", [NSExpression expressionForVariable:@"geometryType"], @[@"LineString", @"Polygon"]];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, expected);
         NSPredicate *predicateAfter = [NSPredicate predicateWithFormat:@"MGL_MATCH($geometryType, 'LineString', YES, 'Polygon', YES, NO) == YES"];
-        auto forwardFilter = [NSPredicate mgl_predicateWithJSONObject:expected].mgl_filter;
+        auto forwardFilter = [NSPredicate predicateWithMGLJSONObject:expected].mgl_filter;
         NSPredicate *forwardPredicateAfter = [NSPredicate mgl_predicateWithFilter:forwardFilter];
         XCTAssertEqualObjects(predicateAfter, forwardPredicateAfter);
     }
@@ -470,7 +470,7 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"{ 6, 5, 4, 3} CONTAINS x"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, expected);
         NSPredicate *predicateAfter = [NSPredicate predicateWithFormat:@"MGL_MATCH(CAST(x, 'NSNumber'), 3, YES, 4, YES, 5, YES, 6, YES, NO) == YES"];
-        auto forwardFilter = [NSPredicate mgl_predicateWithJSONObject:expected].mgl_filter;
+        auto forwardFilter = [NSPredicate predicateWithMGLJSONObject:expected].mgl_filter;
         NSPredicate *forwardPredicateAfter = [NSPredicate mgl_predicateWithFilter:forwardFilter];
         XCTAssertEqualObjects(predicateAfter, forwardPredicateAfter);
     }
@@ -479,7 +479,7 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"%@ CONTAINS %@", @[@"LineString", @"Polygon"], [NSExpression expressionForVariable:@"geometryType"]];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, expected);
         NSPredicate *predicateAfter = [NSPredicate predicateWithFormat:@"MGL_MATCH($geometryType, 'LineString', YES, 'Polygon', YES, NO) == YES"];
-        auto forwardFilter = [NSPredicate mgl_predicateWithJSONObject:expected].mgl_filter;
+        auto forwardFilter = [NSPredicate predicateWithMGLJSONObject:expected].mgl_filter;
         NSPredicate *forwardPredicateAfter = [NSPredicate mgl_predicateWithFilter:forwardFilter];
         XCTAssertEqualObjects(predicateAfter, forwardPredicateAfter);
     }
@@ -488,7 +488,7 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"{ 6, 5, 4, 3} CONTAINS $featureIdentifier"];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, expected);
         NSPredicate *predicateAfter = [NSPredicate predicateWithFormat:@"MGL_MATCH(CAST($featureIdentifier, 'NSNumber'), 3, YES, 4, YES, 5, YES, 6, YES, NO) == YES"];
-        auto forwardFilter = [NSPredicate mgl_predicateWithJSONObject:expected].mgl_filter;
+        auto forwardFilter = [NSPredicate predicateWithMGLJSONObject:expected].mgl_filter;
         NSPredicate *forwardPredicateAfter = [NSPredicate mgl_predicateWithFilter:forwardFilter];
         XCTAssertEqualObjects(predicateAfter, forwardPredicateAfter);
     }
@@ -499,32 +499,32 @@ namespace mbgl {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"a == 'b' AND c == 'd'"];
         NSArray *jsonExpression = @[@"all", @[@"==", @[@"get", @"a"], @"b"], @[@"==",  @[@"get", @"c"], @"d"]];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"a == 'b' OR c == 'd'"];
         NSArray *jsonExpression = @[@"any", @[@"==", @[@"get", @"a"], @"b"], @[@"==",  @[@"get", @"c"], @"d"]];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"NOT(a == 'b' AND c == 'd')"];
         NSArray *jsonExpression = @[@"!", @[@"all", @[@"==", @[@"get", @"a"], @"b"], @[@"==",  @[@"get", @"c"], @"d"]]];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {
         NSPredicate *predicate = [NSPredicate predicateWithFormat:@"NOT(a == 'b' OR c == 'd')"];
         NSArray *jsonExpression = @[@"!", @[@"any", @[@"==", @[@"get", @"a"], @"b"], @[@"==",  @[@"get", @"c"], @"d"]]];
         XCTAssertEqualObjects(predicate.mgl_jsonExpressionObject, jsonExpression);
-        XCTAssertEqualObjects([NSPredicate mgl_predicateWithJSONObject:jsonExpression], predicate);
-        [self testSymmetryWithPredicate:[NSPredicate mgl_predicateWithJSONObject:jsonExpression]
+        XCTAssertEqualObjects([NSPredicate predicateWithMGLJSONObject:jsonExpression], predicate);
+        [self testSymmetryWithPredicate:[NSPredicate predicateWithMGLJSONObject:jsonExpression]
                           mustRoundTrip:NO];
     }
     {